changed yaw gyro ctrl

+++ autopilot/service/util/math/vec2.c
@@ -105,8 +105,8 @@ void vec2_project(vec2_t *vo, const vec2_t *vi1, const vec2_t *vi2)
 
 void vec2_rotate(vec2_t *vo, const vec2_t *vi, float angle)
 {
-   float sa = sinf(angle);
-   float ca = cosf(angle);
+   float sa = sin(angle);
+   float ca = cos(angle);
    vo->x = vi->x * ca - vi->y * sa;
    vo->y = vi->x * sa + vi->y * ca;
 }
